move states up one level to App.js and create a resume page
