+++ src/components/Form.js
@@ -5,59 +5,8 @@ import Education from "./Education";
 import WorkExperience from "./WorkExperience";
 import Projects from "./Projects";
 
-export default function Form(){
+export default function Form(props){
     const [step, setStep] = useState(1);
-    // const [firstName, setFirstName] = useState("");
-    // const [lastName, setLastName] = useState("");
-    // const [portfolioLink, setPortfolioLink] = useState("");
-    // const [city, setCity] = useState("");
-    // const [email, setEmail] = useState("");
-    // const [schoolName, setSchoolName] = useState("");
-    // const [schoolLocation, setSchoolLocation] = useState("");
-    // const [courses, setCourses] = useState("");
-    // const [graduation, setGraduation] = useState("");
-
-    //The three STATES and update functions below could have been combined into one. But the way it is now provides better readibility
-
-    const [values, setValues] = useState({ //holds values for pages Education and Personal details 
-        firstName: "",
-        lastName : "", 
-        email : "",
-        location : "",
-        portfolioLink : "",
-        schoolName : "",
-        schoolLocation : "",
-        courses : "",
-        graduation : ""
-    })
-
-    const [workExp, setWorkExp] = useState({ //holds values for Work Experience (# of workExp are fixed)
-        workPlace1 : "",
-        position1 : "",
-        startEnd1 : "",
-        responsibilities1 : "",
-        workPlace2 : "",
-        position2 : "",
-        startEnd2 : "",
-        responsibilities2 : "",
-        workPlace3 : "",
-        position3 : "",
-        startEnd3 : "",
-        responsibilities3 : ""
-    })
-
-    const [projects, setProjects] = useState({ //holds values for Projects (# of projects are fixed)
-        title1 : "",
-        link1 : "",
-        description1 : "",
-        title2 : "",
-        link2 : "",
-        description2 : "",
-        title3 : "",
-        link3 : "",
-        description3 : ""
-    })
-
 
     function nextStep(){ //increment step value (go to next page)
         setStep(step + 1);
@@ -67,37 +16,14 @@ export default function Form(){
         setStep(step - 1);
     }
 
-    function updateValues(event){ //updates values for Education and Personal Details
-        //console.log(event.target.name, event.target.value)
-        setValues({
-            ...values, 
-            [event.target.name] : event.target.value
-        })
-    }
-
-    function updateWorkExp(event){ //updates values for work Experience
-        setWorkExp({
-            ...workExp, 
-            [event.target.name] : event.target.value
-        })
-    }
-
-    function updateProjects(event){ //updates values for Projects
-        setProjects({
-            ...projects, 
-            [event.target.name] : event.target.value
-        })
-    }
-
-
 
     switch(step){ //switch statement determines which page to go to
         case 1:
             return(
                 <PersonalDetails 
                     nextStep = {nextStep}
-                    values={values}
-                    updateValues = {updateValues}
+                    values={props.values}
+                    updateValues = {props.updateValues}
                 />
             );
         case 2:
@@ -106,8 +32,8 @@ export default function Form(){
                     <Education 
                         nextStep = {nextStep}
                         lastStep = {lastStep}
-                        values={values}
-                        updateValues = {updateValues}
+                        values={props.values}
+                        updateValues = {props.updateValues}
                     />
                 </div>
             );
@@ -117,8 +43,8 @@ export default function Form(){
                     <WorkExperience 
                         nextStep = {nextStep}
                         lastStep = {lastStep}
-                        workExp = {workExp}
-                        updateWorkExp = {updateWorkExp}
+                        workExp = {props.workExp}
+                        updateWorkExp = {props.updateWorkExp}
                     />
                     
                 </div>
@@ -129,8 +55,8 @@ export default function Form(){
                     <Projects 
                         // nextStep = {nextStep}
                         lastStep = {lastStep}
-                        projects = {projects}
-                        updateProjects = {updateProjects}
+                        projects = {props.projects}
+                        updateProjects = {props.updateProjects}
                     />
                 </div>
             );
