+++ src/components/Resume.js
@@ -1,0 +1,11 @@
+import React from "react"
+
+
+export default function Resume(){
+
+    return(
+    <div>
+        <h1>this is the result</h1>
+    </div>
+    )
+}
