--- src/components/Work.js
@@ -1,6 +1,0 @@
-import React from "react"
-
-
-export default function Work(){
-    
-}
